push

+++ app/src/main/java/com/example/yujimomoi/a1daysummerintern/classFile/Point.java
@@ -4,20 +4,20 @@ package com.example.yujimomoi.a1daysummerintern.classFile;
  * Created by yuji.momoi on 2016/07/07.
  */
 public class Point {
-	public float x;
-	public float y;
+	public double x;
+	public double y;
 
 	public Point() {
 		this.x = 0;
 		this.y = 0;
 	}
 
-	Point(float x, float y) {
+	Point(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public void SetPoint(float x,float y) {
+	public void SetPoint(double x,double y) {
 		this.x = x;
 		this.y = y;
 	}
